chore(core): Rename lib/ to portable/
