+++ modules/core/src/portable/compiler-log/format-compiler-log.ts
