+++ modules/core/src/portable/compiler-log/get-shader-info.ts
