+++ modules/core/src/portable/uniforms/uniform-buffer-layout.ts
